requirements.txt
streamlit
yfinance
plotly
pandas
finnhub-python
requests

